feat(M3.5): TimeConveyor-CPU ring — milestone DONE
The TD ring conveyor (дисс. Гл. 2.1–2.2) on z std::jthread nodes with
bounded SPSC channels (core/transport.hpp, CPU stub of ITransport):
node P_k computes pass h = k, k+z, ...; zones stream to the next node
immediately after END; orchestration strictly through ZoneFSM (T1–T5 +
seed §7.1), odd/even I/O order §7.4.

Δt-handoff (B2, дисс. §3.3) via a positional Λ-chain: the j-th sent zone
of pass h carries the full-pass aggregates (v_max/a_max/K2 cap) of pass
h−(n−1−j) — the freshest full pass inside the pass head's light cone —
so dt(h+1) = auto_dt(v_max(h−n+1), dt(h)) is computable at first-send
for ANY node count: the dt sequence is bitwise z-independent (n=1
reproduces the serial rule exactly). INV-4 forecast lag-generalized:
v_pred = v_full + a_full·dt·max(1,n−1). PBC closure via rotation
(ZoneFSM §7.2 [ENG]): head held in c until the tail computes, ENDs and
is sent last; pair set and positions identical to the dissertation's
cross-node packet. Static zone membership + honest StaleZone guard
(migration → M4). Shared per-pair FP core with the serial w-pass
(PairGeom) and shared per-atom reduction kernels (buffer.hpp) — both
refactors bitwise-transparent (config_auto trajectory md5 unchanged).

Tests (8 + NVE): conveyor ≡ serial zone-pass stepper bitwise; 1 node vs
z∈{2..5} bitwise in fixed AND auto modes (per-pass dt equality); Λ-chain
replay oracle from recorded PassStats aggregates + serial-rule oracle at
n=1; PBC closure vs oracle + momentum; PBC+auto; anti-deadlock rings
z=1..6; INV-4/StaleZone/Overlap halts; degenerate inputs (empty zones
under rotation carry the dt header, steps=1, z>steps, 2-zone+PBC ban);
§3.6 replica — Al-72, free boundaries, auto-step, 25 900 steps, 1 vs 4
nodes, zero deviation of coordinates and velocities (рис. 43–44).
Test_NVE_Invariant 50k @ 1 fs/300 K: multizone bitwise ≡ single-zone
(strongest form of "не хуже"); secular trend −1.2e-2 kT/(ns·dof)
[deterministic_fp64, morse+shift] — floor set by the shift-truncation
force discontinuity (аудит C1); ceiling calibrated 3×.

Adversarial review (29 agents: TSan, Debug builds, 564-shape deadlock
matrix; concurrency clean) — fixed: dead-in-Release/false-in-Debug
INV-1 assert (now a computed-flag check + exception → Halt::Internal;
Debug builds green); silent int64 Q24.40 overflow UB on steep
potentials (FixedAccum::add now throws; NaN/inf caught); n=1 dt-rule
lag (decision moved to first send); coincident pairs now trip the
Overlap halt (min_r2 before the acceptance cut); halt-path steps_done/
stats accounting (only provably completed passes); hang on partial
thread spawn; T2 contribution mask implemented + checked at END
(INV-3); accumulator release at SEND; Zone::step_h widened to long.

ZoneFSM doc updated: B2 light-cone lag (§6 п.3–5), rotation closure +
INV-1 carve-out (§7.2), T5 ownership semantics (§4), static membership
(§7.5), rescue deferral and no-assert policy (§9).

M3.5 acceptance criteria green (Roadmap). Next: M4 — CUDA port
(StreamTransport, GPU fixed-point B1, production_mixed, bench baseline).

+++ include/tdmd/core/zone.hpp
@@ -22,7 +22,8 @@ enum class ZoneEvent { RECV, SPHERE, START, END, SEND };
 
 struct Zone {
   int      id = 0;
-  int      step_h = 0;          // time step this zone belongs to
+  long     step_h = 0;          // time step this zone belongs to (long: M3.5
+                                // review — int silently wraps past 2^31 passes)
   int      n_atoms = 0;
   ZoneType type = ZoneType::o;     // free at construction
   double   left_bound_glob = 0.0;  // zone left boundary (for FP32 offsets later)
